+++ src/test/base/RouterSinglePoolWithAssetRecipient.sol
@@ -61,8 +61,8 @@ abstract contract RouterSinglePoolWithAssetRecipient is
         test721.setApprovalForAll(address(router), true);
 
         // Setup pair parameters
-        uint256 delta = 0 ether;
-        uint256 spotPrice = 1 ether;
+        uint128 delta = 0 ether;
+        uint128 spotPrice = 1 ether;
         uint256[] memory sellIDList = new uint256[](numInitialNFTs);
         uint256[] memory buyIDList = new uint256[](numInitialNFTs);
         for (uint256 i = 1; i <= 2 * numInitialNFTs; i++) {
@@ -168,7 +168,7 @@ abstract contract RouterSinglePoolWithAssetRecipient is
             block.timestamp
         );
         assertEq(test721.balanceOf(buyPairRecipient), 1);
-        // Pool should still keep track of the same number of NFTs prior to the swap 
+        // Pool should still keep track of the same number of NFTs prior to the swap
         // because we sent the NFT to the asset recipient (and not the pair)
         uint256 afterBuyPairNFTBalance = (buyPair.getAllHeldIds()).length;
         assertEq(beforeBuyPairNFTBalance, afterBuyPairNFTBalance);
@@ -323,7 +323,7 @@ abstract contract RouterSinglePoolWithAssetRecipient is
             block.timestamp
         );
         assertEq(test721.balanceOf(buyPairRecipient), 5);
-        // Pool should still keep track of the same number of NFTs prior to the swap 
+        // Pool should still keep track of the same number of NFTs prior to the swap
         // because we sent the NFT to the asset recipient (and not the pair)
         uint256 afterBuyPairNFTBalance = (buyPair.getAllHeldIds()).length;
         assertEq(beforeBuyPairNFTBalance, afterBuyPairNFTBalance);
@@ -347,7 +347,7 @@ abstract contract RouterSinglePoolWithAssetRecipient is
             block.timestamp
         );
         // User gets 90% of the tokens (which is output) and the other 10% goes to the factory
-        assertEq(getBalance(address(factory)), output/9);
+        assertEq(getBalance(address(factory)), output / 9);
     }
 
     function test_swapTokenForSingleSpecificNFTWithProtocolFee() public {
@@ -372,8 +372,11 @@ abstract contract RouterSinglePoolWithAssetRecipient is
             inputAmount
         );
         // Assert 90% and 10% split of the buy amount between sellPairRecipient and the factory
-        assertEq(getBalance(address(factory)), inputAmount/11);
-        assertEq(getBalance(sellPairRecipient)+getBalance(address(factory)), inputAmount);
+        assertEq(getBalance(address(factory)), inputAmount / 11);
+        assertEq(
+            getBalance(sellPairRecipient) + getBalance(address(factory)),
+            inputAmount
+        );
     }
 
     function test_swapTokenForSingleAnyNFTWithProtocolFee() public {
@@ -393,7 +396,10 @@ abstract contract RouterSinglePoolWithAssetRecipient is
             inputAmount
         );
         // Assert 90% and 10% split of the buy amount between sellPairRecipient and the factory
-        assertEq(getBalance(address(factory)), inputAmount/11);
-        assertEq(getBalance(sellPairRecipient)+getBalance(address(factory)), inputAmount);
+        assertEq(getBalance(address(factory)), inputAmount / 11);
+        assertEq(
+            getBalance(sellPairRecipient) + getBalance(address(factory)),
+            inputAmount
+        );
     }
 }
