+++ src/test/mixins/Configurable.sol
@@ -16,9 +16,9 @@ abstract contract Configurable {
         ICurve bondingCurve,
         address payable assetRecipient,
         LSSVMPair.PoolType poolType,
-        uint256 delta,
+        uint128 delta,
         uint256 fee,
-        uint256 spotPrice,
+        uint128 spotPrice,
         uint256[] memory _idList,
         uint256 initialTokenBalance,
         address routerAddress /* Yes, this is weird, but due to how we encapsulate state for a Pair's ERC20 token, this is an easy way to set approval for the router.*/
